requirements.txt
feedparser==6.0.11
google-generativeai>=0.8.0
tweepy>=4.14.0
Pillow>=10.0.0
